requirements.txt
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
requests==2.9.1
djangorestframework==3.4.7
trueskill==0.4.4
bonjour-py